+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Layout from './components/Layout/Layout'
+import Layout from './hoc/Layout/Layout'
 import BurgerBuilder from './containers/BurgerBuilder/BurgerBuilder'
 
 //? Components in the 'components' folder are 'dumb' or stateless components
